clv_vscode/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
faiss-cpu>=1.7.4
anthropic>=0.25.0
streamlit>=1.35.0
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
sentence-transformers>=2.2.0
joblib>=1.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
